Refactor the "check categories and severity" logic in diag.t
Move the "todo" bailout as early as possible, before any work is done.
Replace C<use 5.01> [sic] with C<use feature 'state'>, as this makes the
intent much clearer.
Use parenthesis for like() and is() as this is consistent with the rest of the
test, and permits a simplification of the BEGIN handling.