Make t/podcheck.t less sensitive
The max line length is increased to 100.  This eliminates dozens of
known issues.

The warnings about using F<> and L<> instead of C<> have been removed
on the rationale that they are too opinionated.  They flag stylistic
issues rather than syntactic or structural issues and have minimal
impact on the effectiveness of the resulting documentation.