Make \N{  } deprecation warnings fatalizable
What drew my attention to this was the missing category in
perldiag.pod.  I tried adding it, but diag.t complained that it should
be absent.

It thinks it should be absent, because Perl_warn (when used correctly)
does not put the warning in any category, and does not allow it to be
suppressed except via $SIG{__WARN__}.

Use of if(ckWARN) followed by Perl_warn is not correct.  ckWARN checks
to see whether the category is enabled, and then Perl_warn warns with-
out reference to the category at all, so whether it is fatal cannot
be looked up.

The result is that these warnings do not die under ‘use warnings
FATAL => 'deprecated'’.