Update the GSMATCH handling in vms/gen_shrfls.pl.
This code (which only runs if you have set PERLSHR_USE_GSMATCH in
the environment) has not been updated in a long time.  It was
assuming that $] had only five digits after the decimal, whereas
it's had six for some time. And it assumed that the Perl5 version
could be represented in 4 bits, which was true up through 5.15
but isn't true anymore.

So get all the digits of the version number, and go wild and spend
5 bits on the value of $Config{PERL_VERSION}, which will get us
through 5.31.  That only leaves three bits in which to encode all
the options that could break binary compatibility, whereas in fact
we need about thirty bits.

So clearly this only works in a situation where the configuration
can be standardized and/or different configurations are packaged
separately.