Fix ExtUtils::Constant test failure on VMS.
Awaiting upstream application for many months now at:

  https://rt.cpan.org/Public/Bug/Display.html?id=81249

But it's not at all clear that this module is even maintained as
the only two "maintainers" are former pumpkings who probably had
to become maintainers in order to cut a Perl release.

From the message originally submitted upstream:

It turns out the easiest and most robust way to handle the fact that
filename case may or may not be preserved is simply to do all
filename comparisons in a case blind fashion.  This is safe to do
because there is no practical possibility of distinct filenames that
differ only by case, especially with the short list of well-known
files being considered here.