Free operand when encountering unmatched ')' in (?[])
I only need to free the operand (current), not the left-paren token
that turns out not to be a paren (lparen).

For lparen to leak, there would have to be two operands in a row on
the charclass parsing stack, which currently never happens.

(cherry picked from commit 4bc5d08976b7df23b63a56cc017a20ac5766fbbc)